Introducing visitors to interpret the AST
A parsing via pegase generates an abstract syntax tree, which needs
further semantic analysis and interpretation to be performed. This
will be done via visitors, conforming generally to the visitor
pattern, and more speifically to pegase's Visitor type. Currently
one visitor is partially implemented: componentVisitor, which renders
the tree as a JSX.Element.