Multiples of i syntactic sugar
The parser differentiates between '123' and '123i', where the latter
is correctly interpreted as '123 * i'; that is, the imaginary part of
a complex number can be formed as per common convention. This doesn't
seem to impact, per se, the ability to recognize numbers. It might
be strategically useful to extend this to e and pi.