[#4, #5] Functional evaluation
Variables assigned to an expression of other variables can be invoked,
with a special syntax, which pushes temporary values for the variables
used within the invocation on the scope. These values are removed
after the invocation is complete. In the event that an undefined
variable is used, the function body is rewritten using that undefined
variable, and is the value of the assignment. Parameters are assigned
values from the supplied arguments based on index, left-to-right.
Excess arguments are ignored; if fewer arguments are passed than
parameters, the result is a new expression in the remaining undefined
variables.