[#7] Implementing digamma: UX and parser
Keypad can generate the psi symbol, which the parser interprets as
digamma. Full evaluation is forthcoming, but most of the visitors
will accept DIGAMMA labelled nodes.

Note that evaluateVisitor now uses closures for unary and binary
operations. This might be slightly less efficient than the previous
implementation, as copies of the underlying functions are likely
being made for each closure; however, it is less code, which is nice.