[#7] Skeletal digamma; isInteger
Fields now abstractly provide a method, isInteger, to ascertain if
the stored value is an integer. Real and complex implement.

Field also has a skeleton for digamma, which will have about five
different cases that need implementation. One of those cases, the
half integers, will necessitate the implementation of a guard
function abstractly on fields to ascertain that statis.