simplifyVisitor, for symbolic reductions
EXPONENT is now RAISE, for internal consistency
simplifyVisitor attempts to apply a number of rules at various nodes,
where any given rule that matches results in a semantically
equivalent but syntactically cheaper tree. The ideal is the reduce
the total number of nodes to represent then same operation, while
also generating output that is easier to interpret.