Merged the old field classes into the new ones
Note: this showcased a circular dependency with Visitor; while the
treeParser spec worked without issue, the specs for Real and Complex
did not. As an experimental test, Visitor declare vague information
about the types it uses to define its interface, thus absolving it of
needing to import each type. This seems to work (the broken specs run)
but I cannot verify the integrity until testing of concrete visitors
is undertaken.