Experimenting with a potential Multi-Visitor
PairVisitor defines an interface for a potential class to implement.
The methods suggested by the interface take the form of
, with the appropriate types for T and U listed as
arguments of the method. These functions are optional, so an
implementation can skip unnecessary permutations. This necessitates
the existence of a required 'visitDefault'. The optional methods
would for specific typings for the parameters and return values, so
an implementation couldn't do something like
visitRealAndVariable(a: Addition, b: Complex): never.

Allowing multi-dispatch on this is complicated: in theory, a given
invoking context would simply need to call an 'accept' method on the
first object, pass the visitor and the second object, and get routed
to the appropriate visit overload. However, this requires that each
Node type implement these accept methods, which requirs knowledge of
a broader set of other objects. A bit messy.

Motivation: make the pattern matching in Simplification less stupid:
each operator could be redefined as a PairVisitor, which is delegated
to.