Experimentations with a functional approach
As a though experiement, an alternative approach to handling tree,
evaluation, and simplifiication has been explored. The core idea is
that a full OOP system design trends towards cumbersome pedantry;
while the extant system is much more type safe than the loose typing
of pegase's AST and Visitors, it is also ridiculously more verbose.
The skeleton of a functional approach here is different: minimal
object hierarchy (which might be eliminated in favor of typed POJOs),
with visitors replaced with a discrete functional suite. For example,
the numeric and algebraic behaviors of Evaluation and Simplification
for Addition nodes has been (mostly, and incompletely) folded into
the provided add function.

Functions are either singularly valued, as with real, multimethods
which handle a variety of different edge cases, or special decorators
or partial evaluations of another function. The latter cases are
interesting: division, for example, is a function which wraps
multiply by providing a reciprocal of its second argument as the
second argument to multiply. This trends towards minimizing the
number of edge cases which need be considered for algebraic
reduction.