Arcus functions have been implemented
This required a whole lot of special cases for a variety of other
operations to properly handle the intersection of reals and complex
numbers. Some of this might be obviated by having 'real' and 'complex'
return the opposite types in certain edge cases. Doing so might cause
certain functions which expect to be receiving a complex to get a
real, which might be problematic.