The Grand Yeeting
Most of the vestiges of the pedantic OOP approach which preceded this
variant have been removed. The remainder (i.e. a couple of visitors)
still exist in a redacted form as they have functionality which still
needs porting.

The treeParser should now be correctly operating on the new
functional approach. The UI is utilizing the proper rendering
context. Functional nearly done.