Variable assignement and dereferencing
If the parser has scoped context, it will attempt to store bound
variables withint hat scope when assignment expresssion are
encountered. The value of an assignment is the value assigned to the
variable. When a variable is referenced, the parser attempts to
retrieve it from scope; if the variable is bound, the value of the
bound variable is returned; otherwise, an unbound variable is
returned. Bound variables do not normally exist within the AST.