Simplified binary generator; unaryFrom
Binary operations likely always want the real/complex manifold to
delegate to the complex/complex case, so this is automatically
done in the closure generator.

unaryFrom attempts to generate a new derived unary function from
a binary function generated from binary. Should be capable of
specifying which child the provided value is bound to.