Polygamma mostly complete
digamma uses both recurrence and reflection formulae to calculate a
value for a large input related to a small or negative number.

polygamma is doing the same; the recurrence relation works correctly,
afaic, but the reflection does not currently: to implement it, the
ability to differentiate a function, and then invoke it with a value
is required. Differentiation is not necessarily an issue, but the
invocation logic is not currently implemented. One note: Polygamma,
the class, needs to be defined before differentiate, the function,
is imported in the polygamma.ts file. circular references galore.

Likely will need to special case the definition of Polygamma.
To the land of misfit objects.