Altered differentiate to allow polygamma to use
polygamma would benefit from being able to calculate a derivative, but
would generate an error if it attempted to include differentiate, as
the latter require Polygamma to be defined. By changing how
differentiate type checks Polygamma, the appropriate dispatch still
happens for that derivative, and polygamma can make use of
differentiate.