Invocation logic complete
Could probably use a looking at when less tired. Likely could change
the way that the evaluate logic is being generated so as to avoid
having to (potentially) recreate it each invoke.

Nevertheless, invoke works as a construct within the code. Still
needs to be wired into the parser.

Side benefit: polygamma's reflection formula now works.