is extracted to separate file; bug fixes galore
predicates requires equals, which requires several expression classes,
which was causing circular references. Moving is to a separate file
allows the expression classes to use it to define themselves without
impacting predicates or equals.

Invocations were borked; for some reason, the custom dispatch was not
playing well with the predicate checks. So, it is back to being a
closure generator. Slower, perhaps, but works.