valuesBetween, a bounded generator concept
Specifies a range of values, [min, max), which are split up into
slices increments. Each step, starting from min, is passed as an
argument to an invocation of the provided expression.

Graph should be able to use this--or some varient thereof for 3d
functions--to generate position tuples for generating spline segments.