Graphing Proof of Concept
While incredibly rudimentary, and lacking in QoL features, a proof
of concept implementation of graphing of functions of 1 variable is
now prosent. When an expression is entered, a button is rendered next
to the expression which adds it to the current graph.

Graph delegates to an internal component, Curve, which invokes over
the expression numerous times to generate line segments. None of this
is currently context sensitive, so the process is very slow. Also,
no distinctive colors are rendered for multiple functions, nor any
labels, so the output is a mess. But it works.