Brought unary in line with binary
unary was not designed to have a similar functional footprint to the
conceptually similar HoF binary; this has been rectified: unary now
requires a type constructor function as parameter, which it uses to
generate a fallback case. Generated unaries now only need to specify
type-specific edge cases. Note that digamma required a separate HoF
to make this system work.