Functionally defined keys
The original skeuomorphic design of the keypad left some design
atavisms in the app; specifically, keys were cumbersome to define,
in part as they had three distinct states which the app was attempting
to pedantically type-check enforce. This become unwieldly with the
alpha split and trig modes.

New hotness attempts to rectify that by introducing a slightly less
pedantic set of functions: createKey for generating Key nodes, and
createMode for creating generator functions for various ModeProps.

The result, while a bit of a React abomination, is much more succinct.
Individual keys can specify exactly which modes they support, and edge
cases are both exceptional and minimally specified.