Unary now supports bools. Logical Complement.
unary will now coerce boolean values to either real 1 or 0. Any given
unary function can alternatively override the coercion with custom
behavior when encountering a boolean.

logicalComplement provides not, a function for providing the logical
complement of the argument. For the most part, this is used for
boolean values to generate the opposite of the input.