NotEquals integrated into the parser.
As noted in a comment in parser.ts, there is an awkward clash between
factorial and not equals, where the bangs used in both can lead to
the grammar not being able to resolve not equals. A work around to
address this might end up breaking certain edge cases wrt equals.