Alterations to symbols, front-end edition
While not yet implemented in the parser, the front-end now has
slightly modified symbols for certain behaviors, to help avoid
clashes on new logic operators and to help disambiguate two distinct
operations.

The most important change is altering assignment to use ':=' instead
of '<-', so the latter symbol can be used without clash for converse.