binary function specification simplification
'binary' now requires one extra functional invocation to generate the
multimethod it constructs; this latter function allows extra methods
to be prepended to the multimethod, in a fasion similar to fromMulti.

This was done to simplify usage behavior: instead of capturing the
result of 'binary', having to typeof it, and then export a typed
version of fromMulti, this is consolidated in a single invocation
chain. Much simpler, all things considered.