Potential candidate for extensions
While necessitating a removal of the dispatch function, 'binary' now
has behavior that allows it to handle special extension methods which
encapsulate Writer behavior. Specically, 'when' will generate both
the binary array for the case value to method and handle Action
results from a corresponding function. This mostly simplifies the
overall design of dealing with extension methods: any given
edge case, should it be written with 'when', will automatically
handle the logging information, albeit in a way that passes around
the writers arguments. This latter behavior is needed for edge cases
where other AST-HOF-closures need to be invoked on the parameters to
get the resulting value---which is most extant edge cases.