Restructured when to be more 'method'-compliant
While maintaining type safety and ease-of-typing, 'when' is now more
idiomatically of the function signature of 'method', rather than the
previous partial-eval/currying scheme. This allows, amongst other
things, arbitrary predicates to be supplied to it.

An example of the new behavior can be seen with the 'combine
primitives across nesting levels' edge case, which also provides an
interesting approach to handling typed children within a tree.