The binary closure should now handle return types
Like the broader app, 'binary' will now allow invoking contexts to
specify an explicit return type which will be enforced for most of
the static cases it provides. NaN and Nil will still generate NaN.

A brief, incomplete implementation of equals is provided to showcase
this functionality.

Should an invoker not request a return type, each edge case will have
a default that is contextually intuitive (e.g. Real to Real).