Implemented deepEquals for equality checks
This function attempts to assert that two Writer monad values are
deeply equal to each other, traversing their trees to verify. Takes
into account node's kinds as well as values. Specially handles NaN
as never equivalent to itself.

The function adheres to the BinaryPredicate<T> type, so it can be
directly used within 'when' as the latter's predicate function.

A basic 'multiply' is added to facillitate testing of this new
function