Added 'deepEqualsAt' for targeted checks
deepEqualsAt wraps deepEquals: its parameters tell it which pathways
of the predicate arguments should be sent to deepEquals. This allows
a binary, for example, to walk into either its left or right subtree
for the comparison, rather than against the binary itself.

Combined with 'when', 'deepEqualsAt' should replace the previous
misnomed 'visit' function.

An example use is provided on add for combining terms across a nesting
level.

Note, there are some peculiarities of visit that this does not yet
fully account for, as seen in the current implementation of add.
Further investigation is needed.