Further dissemination of writer doodle
1. unary and supporting types and functions have been moved to the
closure file.
2. Example unary functions have been moved into their forever homes.
3. binary and its supporting types and funcitons mostly moved over
to the closure file.
4. add moved to its forever home, but is currently incomplete. Need
to revise deepEquals and move multiply to finish.
5. Revised AST node definitions to be more potently extensible:
 a. Base type is TreeNode, which defines 'species', 'genus' and
'clade'. These collectively categorize a node. A series of closure
generators for predicates to type guard those values now exist.
 b. Species, an enum, denote non-abstract nodes, like Real or
Addition. The enum values need to be kept up, but should minimize
cyclic dependencies.
 c. Genera, an enum, denotes thematic groupings of functions, such
as the trigonometric set. An 'isGenus' will be true if a node's
genus field matches one of these enum values, and will allow the
node to be treated as an abstract node.
 d. Clades, an enum, is largely for arity. While genus can be
undefined, everything has a clade. An 'isClade' will be true as
long as the node's clade is equivalent.
6. With these revisions, it becomes easy to define abstract nodes
which can broaden the scope of an object, making handling an
abstraction like primitives easier.
7. Should also obviate the need to explicitly create giant union
types for linked node types.