Efforts made to unravel multiply edge cases
Not successful at the moment. Attempting to eliminate need for
'canFormExponential' and 'exponentialCollect', as those functions
are both ridiculously recursive and convoluted. They are hard to
reason about and extend. Unraveling their logic and incorporating
into 'multiply' makes the latter function more complicated, but does
so in a way that is easier to reason about. WIP.