Invocation migration, wip; partially functional
While incomplete---it lacks coverage of quite a few functions---the
'invoke' function has been mostly migrated to the new monadic design.

It operates on Writer<TreeNode> objects, but with a twist: it wraps
those objects in the new Reader monad, bundling a Scope object as
an environment that can be used to contextualize the evaluation.

Crucially, this means the 'evaluate' multimethod does not need to be
rebuilt for each invocation to inject the scope dependency into it,
which should be slightly more performance friendly.

Meanwhile, the edge-cases in 'evaluate' are minimally different from
the previous version. In most cases, the 'binary' and 'unary' helper
functions encapsulate most of the changes, and a newly introduced
'when' function takes care of creating the method instances and
ensuring the monadic bindings.

Future commits will finish off the migration, which should be
comparably trivial.