Move MathSymbols and stringify. Fixed references
Migration of client continues, with a few bits of its old bundled
library migrated to frontend and calcula. Meanwhile, a number of
files have been updated to reflect where things are now located, and
to account for changes in object structure. Notably, stringify now
should work with the new design present in calcula.