Changes to logging and derivation output
1. Writer now stores a singular input (the node effectively being
created) and an output (the result of the operation) in each logged
operation.
2. The closures no longer expose their 'when' function, as this now
requires the generated 'create' function to generate the new 'input'
within the logged operation. Instead, 'when' is accessed as a
parameter on an edge case callback. Not entirely clean, but clever.
3. Most operations have been updated to reflect these changes.
However, none of the test suites in calcula have been updated to
reflect changes to log, so a bunch of errors which need rsolving.
4. Derivation has been updated to generate more sensible (-ish)
output reflecting the above changes. It now generates a grided-div
for its output, with each row representing a step undertaken.
Through clever use of CSS counters, ::before content, and grid-
alignments, the steps are pretty clean and read nicely. There
might be some structural redundancies which need cleaning, and
some log information may be out of order. Will need further
investigation.