Inequality spec updated to changes to writer
Addition partially updated, but inconsistencies in output have put a
pause on completion. Notably, the approach for designating the output
of any given step in the log can lead to apparent out-of-order
deductive leaps. While some of this can be laid at the feet of
Expression, most of it is because of the recursive nature of the
app: the output for a step that delegates to another function is
naturally going to be a final result, rather than an interstitial.

A potential solution, which needs analysis prior to committing to yet
another round of edits, would involve re-imbuing the 'when's with
the create functions for the intention of using them to generate
the output of a given operation. This could work, but would likely
pose some issues. (E.g., what if a tree rewrite delegates to a
function that is outside the context of the operating function? E.g.
adding two like things rewrites to a doubling.)

Alternatively, 'Action' could be updated to require an explicit
output, which could be filled with the create hooks. This is terribly
redundant and bulky, but would better signal intent.