Potential solution to logging intentionality
To ensure that any given operation in calcula logs a meaningful
intended transformation, this commit explores a potential solution:
explicitly detailing the form of the operation as part of the Action
tuple return value given to the when functions.

This rewrite output is intended to depict the form of expression that
is being generated, which is largely lost on the stack.

This is, potentially, facilitated by a newly introduced function,
'rule', which acts as a tag function to process a template literal
representation. While the development of this stuff needs a
consistency pass, the process seems sound: 'rule' generates a nullary
closure which, upon invocation, generates the string representation
of the intended output of the associated step.

(The nullary closure is to avoid unnecessary string processing on
certainly teim-constrained processes, like 'invoke'.)

Problem: 'rule' currently depends upon 'stringify', which is, I
believe, introducing some circular references which TS is unable
to untangle. This is causing specific operations to report that
some functions---such as 'binary'---are not functions. Will need
to evaluate a possible redesign of rule's dependencies.