Dependency injection to resolve cyclical code
Stringify was causing issues with functions not being able to invoke
one another before being properly defined. By removing it as a direct
dependency of 'rule', and changing the nullary closure generated by
that function to instead by a unary that accepts a stringify function
as a parameter, the cyclic dependencies are gone. This also allows
for the function which generates a string output to ultimately be
configurable. Double win.