Added rules for most of differentiate
Arcus and AreaHyperbolic not yet covered. Likely an easier way to
go about doing this. Possibly by altering 'rule' to accept other
functions that adhere to a specific signature, or to a tuple that
encapsulates the parameter list from another tag. This would allow
for sub-rules which could simplify common patterns (like chained
derivatives, for example).