Partial success wrt output. Input needs fixing.
'differentiate' and 'invoke' updated---ish---to new 'rule' system;
'Derivation' updated to make use of this new design. Generated output
is, for the most part, much better; it captures the intent of a given
step more than adequately.

However, input is a bit nonsensical, especially for derivatives. The
process of fixing that would require putting it through the same
approach that is being used to generate the output. However, this can
me automated, I suspect painlessly. The 'when' functions would need
updating to make use of a newly auto-generated closure rule, as well
as a create function. Unlike the last attempt at so infusing the
'when's, the new approach would use dependency injection: each 'when'
would generate a function which the closure hofs would have to
invoke with the appropriate context to create the actual multimethod.

Further adjustments are needed to the primitive handlers within the
hof closures. Specificially, there is a lot of redundancy which might
be simplified. (E.g., many operations do the exact same process for
all three primitive types, or could handle the boolen case with a
type-cast behind the scenes to real.) Worth taking a look at.

Regardless, the primitive functions could be simplified to just
return the new value, and their Action tuple could be built by the
hof closure instead. Less work all around.