Logged inputs now input, a Rewrite
The various 'when' functions now create a Rewrite for the new 'input'
field or the generated log. For 'binary' and 'unary' derived functions,
'when' requires a 'toString' implementation passed to it to provide
this functionality. This is handled behind the scenes by these
respective functions, which allows most derived closures to not have
to be aware of the functionality. However, the closures needed
updating to reflect new data they needed to provide to their HOFs:
a 'name' string and a Notation, which are collectively used to
generate the Rewrite.

The input Rewrite should eventual take the place of the ad hoc *Rule
functions which were previously used; this will follow in the near
future. 'binary' and 'unary' need a cleaning on their primitive
handlers to make the process of defining new functions less onerous.
Meanwhile, 'primitive' is a mess that needs deep cleaning.