Trigonometric function specs updated
In addition to making the specs for the trig functions correctly
reflect the current behavior of 'unary', the latter function has also
been updated to better log intent for certain rewrite-heavy edge-
cases. This new functionality allows an invoking context of 'unary'
to deliberately specify the rewrite function which will be used in
the primitive handlers, rather than use the default. This is used,
for example, in 'sec' to signal the intent of the operation is to
delegate calculation to 'reciprocal' and 'cos'.