Introduced rules  and
Both rules are meant to take the place of a standard use of ;
each inserts a caret into the rendered description to denote the
location where processing of the implied expression is at that point.

 prepends the caret, indicating that further processing
needs to be conducted to reach completion. By contrast,
places the caret at the end of the description, indicating that
the preceeding result has been fully processed.

Different functions, in different edge cases, will make use of both
of these rule variants to indicate what operations have been done.