Terminal layout tweaks; inequality logging fix
1. Terminal has changed its grid to a two track layout. Each track
contains a container element: the first for the marker for that entry;
the second for the entry details. This setup allows the entry markers
and controls to sticky properly, with a visual push off screen when
their respective containers are no longer visiable. Various rendering
issues related to operators have also been addressed.

2. equals and notEquals were not using the proper operators for their
logging. This has been updated to change the incorrect '===' and '!=='
to the correct '==' and '!='.