Yet Another Test Suite Rewrite
The whole Ops suite to verify writer logs was very messy and tedius.
This commit features the introduction of a new expectation,
'expectToEqualWithSnapshot', which makes use of jest's snapshotting
feature to keep records of known values of logs that are then
compared against, automatically, in the future.

This is not without some potential pitfalls: snapshots are a bit
chonky (even after a custom serializer was introduced), and they
require a review step to verify that successive tests are validating
against reasonable reference output.

Still, this will likely obviate the need for the Ops functions and,
in the process, slimify most of the test suites directly, just so
long as no one glances into the __snapshot__ directories.