[wip] 'multiply' and 'add' rewrite
The above two functions will soon be reimplemented as a new node type:
'MultiaryNode', which is generated by 'multiary'. This funciton is
similar in concept to 'binary' and 'unary', but stores all of its
operands as a single list, upon which various set operations are
performed to conduct the needed computation.

'multiary' has been designed to reduce the overall complexity of
'multiply' and 'add'. As such, the latter two functions are deeply
pared down compared to their previous implementations. (Note that
'add' is incomplete; neither have handlers for specific numeric
operands values like 0 or infinity.)