Special numbers in multiary functions
'multiary' now supports a replacement block, which takes edge cases
which should alter the operand list if the first item in the list
is a special number. 'add' and 'multiply' have been updated to use
this.

The tail end of the file also has some exploratory code for the
upcoming update to 'primitive' to allow derived closures to have
better argument lists. Weird but nifty behavior going on with that
logic.