[#17] Raw primitives and generic rest params
Primitive closures now take generic rest parameters, allowing them to
be both typesafe and easier to work with. Of primary note, 'complex'
now takes two parameters, the number values for 'a' and 'b', rather
than a tuple. This required many changes downstream.

PrimitiveNode's now have a dedicated 'raw' field (except Nil), which
is either the type of the raw system data type, or is an object of
same. For Real and Boolean, this is, respectively, 'number' and
'boolean'; for Complex, this is '{a: number, b: number}.
Consistency, ho. This required many changes downstream.