[#17] While a bit ugly, primitive extensions
Due to the weirdness of generic rest parameters, 'when' needs to be
generated by the 'primitive' hof and exposed to the closure via a
parameter to the extensions block. This is a bit messier than how
the other hofs handle their 'when' functions, but it ensures that the
primitive when's have the appropriate rest params typings.