[#18] MultiaryNode operands sorted
'multiary' will now automatically sort its operands within the
generated 'create' function by degree. A closure generated by
'multiary' will need to supply to the hof a sort order, which will
cause 'create' to reverse the sorted operands if descending.

Sorting is handled by a function in multiary.ts: 'sortBy', which
takes an array to sort and a mutation method. The mutate method is
applied to each element of the array to convert those elements to
raw numbers. These numbers then use a standard comparator for
sorting. 'sortBy' minimizes the calls to 'mutate' by mapping its
input first, and sorting the map. This could be inefficient for
large arrays, as extra memory would be needed to keep things around.