[#18] wip. Updating tests
Working on migrating the old binary definitions for add and multiply
to the new multiary implementation. Many test suites are working,
but there is a lot of behavior that is currently not working. Will
need to figure out how to get behavior like 'degree' into multiary;
currently, this causes issues (e.g. 'degree' gets invoked before
'multiary' is fully defined).