[#18][wip] Introduction of otehrwise
Further attempts made to fix the issues with the bugs in multiary.
There is some excess logging happening in certain edge cases, and
not enough logging in others. Likely the behavior defined in
otherwise needs to be combined, in some fashion, with the callback
passed to it, though that might be messy. Needs massive thought.