Alternate universe writer implementation
Designing a potential alternative to the current implementation of
monadic, that focuses on attempting to make writers work more
idiomatically. Notably, this should keep writers out of the TreeNodes,
but allow for bare TreeNodes to be passed into functions which
operate with Writer wrapped inputs.

Also of note: basic design of a diff-based log system. Needs a lot
more thought to be of use.