Some notes on unaryBind
The variant presented yesterday was incompatible with NuLog. A
slight alteration in this commit is still incompatible, but is a
step in the right direction, albeit in a way that is a bit onerous.
Need to think about how the logging should actually be done if this
whatIf scenario is to be persued further.